rahul/preprocessing/requirements.txt
# RadStream Lambda Functions Requirements
# Core AWS and Python libraries
boto3>=1.34.0
botocore>=1.34.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# JSON handling (built-in, but explicit for clarity)
# json - built-in module

# Date/time handling (built-in, but explicit for clarity)
# datetime - built-in module

# Base64 encoding (built-in, but explicit for clarity)
# base64 - built-in module

# IO operations (built-in, but explicit for clarity)
# io - built-in module

# OS operations (built-in, but explicit for clarity)
# os - built-in module

# Time operations (built-in, but explicit for clarity)
# time - built-in module

# Type hints (built-in in Python 3.5+)
# typing - built-in module

# Note: The following are AWS Lambda runtime dependencies that are pre-installed:
# - json
# - datetime
# - base64
# - io
# - os
# - time
# - typing
# - boto3 (with specific version)
# - botocore (with specific version)

# For local development and testing, you may also want:
# pytest>=7.0.0
# moto>=4.0.0  # For mocking AWS services
